refactor: place dto into folders
