+++ Recademy.Api/Controllers/ProjectController.cs
@@ -2,7 +2,7 @@
 using System.ComponentModel.DataAnnotations;
 using Microsoft.AspNetCore.Mvc;
 using Recademy.Api.Services.Abstraction;
-using Recademy.Shared.Dtos;
+using Recademy.Shared.Dtos.Projects;
 
 namespace Recademy.Api.Controllers
 {
