+++ Recademy.Api/Services/Abstraction/IAchievementService.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
-using Recademy.Shared.Dtos;
 using Recademy.Core.Models;
+using Recademy.Shared.Dtos.Achievements;
 
 namespace Recademy.Api.Services.Abstraction
 {
