+++ Recademy.Api/Services/Abstraction/IReviewService.cs
@@ -1,5 +1,5 @@
 ﻿using System.Collections.Generic;
-using Recademy.Shared.Dtos;
+using Recademy.Shared.Dtos.Reviews;
 
 namespace Recademy.Api.Services.Abstraction
 {
