+++ Recademy.Api/Services/Implementations/AchievementService.cs
@@ -1,8 +1,8 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using Recademy.Api.Services.Abstraction;
-using Recademy.Shared.Dtos;
 using Recademy.Core.Models;
+using Recademy.Shared.Dtos.Achievements;
 
 namespace Recademy.Api.Services.Implementations
 {
