+++ Recademy.Shared/Dtos/Github/GitHubIssueCreateDto.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace Recademy.Shared.Dtos
+namespace Recademy.Shared.Dtos.Github
 {
     public class GitHubIssueCreateDto
     {
