+++ Recademy.Shared/Dtos/Github/GithubProfileDto.cs
@@ -1,4 +1,4 @@
-﻿namespace Recademy.Shared.Dtos
+﻿namespace Recademy.Shared.Dtos.Github
 {
     public class GithubProfileDto
     {
