+++ Recademy.Shared/Dtos/Reviews/RequestsByFilterDto.cs
@@ -1,4 +1,4 @@
-﻿namespace Recademy.Shared.Dtos
+﻿namespace Recademy.Shared.Dtos.Reviews
 {
     public class ReviewRequestSearchContextDto
     {
