+++ Recademy.Shared/Dtos/Reviews/ReviewProjectDto.cs
@@ -1,4 +1,4 @@
-﻿namespace Recademy.Shared.Dtos
+﻿namespace Recademy.Shared.Dtos.Reviews
 {
     public class ReviewProjectDto
     {
