+++ Recademy.Test/Controllers/ProjectControllerTests.cs
@@ -1,9 +1,10 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using NUnit.Framework;
-using Recademy.Shared.Dtos;
 using Recademy.Mock.Generators;
 using Recademy.Test.Tools;
+using Recademy.Shared.Dtos.Projects;
+using Recademy.Shared.Dtos.Users;
 
 namespace Recademy.Test.Controllers
 {
