+++ Recademy.WebUi/Pages/Achievements.razor
@@ -1,5 +1,5 @@
 ﻿@page "/achievements"
-@using Recademy.Shared.Dtos;
+@using Recademy.Shared.Dtos.Achievements
 @inject HttpClient Http
 
 <h1>Достижения</h1>
