+++ Recademy.WebUi/Shared/MainLayout.razor
@@ -1,4 +1,4 @@
-﻿@using Recademy.Shared.Dtos;
+﻿@using Recademy.Shared.Dtos.Users
 @inherits LayoutComponentBase
 @inject ILogger<MainLayout> Logger
 @inject HttpClient Http
